% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2.R
\name{cor2}
\alias{cor2}
\title{Correlaciones pairwise}
\usage{
cor2(data, round = 2)
}
\arguments{
\item{data}{data frame}

\item{round}{redondear, por default 2 decimales (round, by default 2 digits)}
}
\value{
matriz de correlaciones (matrix correlations)
}
\description{
Correlaciones pairwise
}
\examples{
cor2(mtcars)
}
