% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seguridad.R
\name{crear_grafo_desde_tabla_enlaces}
\alias{crear_grafo_desde_tabla_enlaces}
\title{Crear grafo a partir de lista de enlaces.}
\usage{
crear_grafo_desde_tabla_enlaces(fichero, identNumerico, sep = "")
}
\arguments{
\item{fichero}{ruta al archivo que contiene la lista de enlaces, puede ser .txt o .csv.
Cada fila del fichero contiene los dos vértices que forman cada enlace.}

\item{sep}{carácter que separa en el fichero los dos vértices de cada enlace. Por defecto será
uno o más espacios en blanco}

\item{identificadores}{booleano que indica si el identificador de cada vértice es númerico
o no. En caso de ser númerico la función lo convierte a carácteres.}
}
\value{
el grafo correspondiente a la lista de enlaces.
}
\description{
Función que genera un grafo a a partir de la lista de enlaces entre vértices.
}
