% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seguridad.R
\name{crear_subgrafo_estrella}
\alias{crear_subgrafo_estrella}
\title{Subgrafo estrella.}
\usage{
crear_subgrafo_estrella(grafo, lonMuestra)
}

\arguments{
\item{grafo}{grafo que se quiere muestrear .}

\item{lonMuestra}{tamaño del conjunto inicial de vértices de la muestra.}
}
\value{
subgrafo muestreado.
}
\description{
Función que crea una muestra de un grafo con la técnica de muestreo estrella.
}
