% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seguridad.R
\name{crear_subgrafo_incidente}
\alias{crear_subgrafo_incidente}
\title{Subgrafo incidente.}
\usage{
crear_subgrafo_incidente(grafo, lonMuestra)
}
\arguments{
\item{grafo}{grafo que se quiere muestrear .}

\item{lonMuestra}{tamaño del conjunto inicial de enlaces de la muestra.}
}
\value{
subgrafo muestreado.
}
\description{
Función que crea una muestra de un grafo con la técnica de muestreo incidente.
}
