% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seguridad.R
\name{estimacion_grado_medio_con_subgrafo_estrella}
\alias{estimacion_grado_medio_con_subgrafo_estrella}
\title{Estimador del grado medio del grafo con muestreo estrella}
\usage{
estimacion_grado_medio_con_subgrafo_estrella(
  estrella,
  numVertices,
  distribuccionesEst,
  lonMuestra0
)
}
\arguments{
\item{estrella}{subgrafo generado mediante muestreo estrella.}

\item{numVertices}{número del vértices del grafo original.}

\item{distribuccionesEst}{lista con los grados de los vértices muestreados.}

\item{lonMuestra0}{tamaño del conjunto inicial de vértices utilizado
para el muestreo estrella.}
}
\value{
estimación del grado medio de los vértices del grafo.
}
\description{
Estimador del grado medio del grafo. Usa una muestra del grafo
generada mediante muestreo estrella y el estimador de Hortvitz-Thompson.
}
