% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seguridad.R
\name{estimacion_num_enlaces_con_subgrafo_estrella}
\alias{estimacion_num_enlaces_con_subgrafo_estrella}
\title{Estimador del número de enlaces con muestreo estrella.}
\usage{
estimacion_num_enlaces_con_subgrafo_estrella(
  estrella,
  numVertices,
  lonMuestra0
)
}
\arguments{
\item{estrella}{subgrafo generado mediante muestreo estrella.}

\item{numVertices}{número del vértices del grafo original.}

\item{lonMuestra0:}{tamaño del conjunto inicial de vértices utilizado
para el muestreo estrella.}
}
\value{
estimación del número de enlaces.
}
\description{
Estimador del número de enlaces de un grafo. Usa una muestra del grafo
generada mediante muestreo estrella y el estimador de Hortvitz-Thompson.
}
