% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seguridad.R
\name{estimacion_num_enlaces_con_subgrafo_inducido}
\alias{estimacion_num_enlaces_con_subgrafo_inducido}
\title{Estimador del número de enlaces con muestreo inducido.}
\usage{
estimacion_num_enlaces_con_subgrafo_inducido(inducido, numVertices)
}
\arguments{
\item{inducido}{subgrafo generado mediante muestreo inducido.}

\item{numVertices}{número del vértices del grafo original.}
}
\value{
estimación del número de enlaces.
}
\description{
Estimador del número de enlaces de un grafo. Usa una muestra del grafo
generada mediante muestreo inducido y el estimador de Hortvitz-Thompson.
}
