% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seguridad.R
\name{estimacion_num_vertices_cap_recaptura}
\alias{estimacion_num_vertices_cap_recaptura}
\title{Estimador del número de vértices del grafo con captura-recaptura}
\usage{
estimacion_num_vertices_cap_recaptura(grafo, lonMuestra1, lonMuestra2)
}
\arguments{
\item{grafo}{grafo del que se quiere calcular el número de vértices.}

\item{lonMuestra1}{longitud del primer conjunto de vértices muestrado.}

\item{lonMuestra2}{longitud del segundo conjunto de vértices muestreado.}
}
\value{
estimación del número de vértices del grafo.
}
\description{
Estimador del número de vértices del grafo usando la
técnica de captura-recaptura.
}
