% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seguridad.R
\name{estimacion_num_vertices_con_subgrafo_incidente}
\alias{estimacion_num_vertices_con_subgrafo_incidente}
\title{Estimador del número de vértices del grafo con muestreo incidente}
\usage{
estimacion_num_vertices_con_subgrafo_incidente(
  incidente,
  numEnlaces,
  distribuccionesInc
)
}
\arguments{
\item{incidente}{subgrafo generado mediante muestreo incidente.}

\item{numEnlaces}{número de enlaces del grafo original.}

\item{distribuccionesInc}{lista con los grados de los vértices muestreados.}
}
\value{
estimación del número de vértices del grafo.
}
\description{
Estimador del número de vértices del grafo. Usa una muestra del grafo
generada mediante muestreo incidente y el estimador de  Hortvitz-Thompson.
}
