% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seguridad.R
\name{estimacion_num_vertices_estrella_sin_HT}
\alias{estimacion_num_vertices_estrella_sin_HT}
\title{Estimador del número de vértices del grafo con muestreo estrella.}
\usage{
estimacion_num_vertices_estrella_sin_HT(grafo, p0)
}
\arguments{
\item{grafo}{grafo del que se quiere calcular el número de vértices .}

\item{p0}{probabilidad de inclusión de cada vértice en la muestra inicial.}
}
\value{
estimación del número de vértices del grafo.
}
\description{
Estimador del número de vértices del grafo. Usa la técnica de estimación
del tamaño de poblaciones con muestreo estrella, que no utiliza el estimador de Hortvitz-Thompson.
}
