% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_a_centigrados.R
\name{fahrenheit_a_centigrados}
\alias{fahrenheit_a_centigrados}
\title{Funcion para covertir grados fahrenheit a centigrados}
\usage{
fahrenheit_a_centigrados(temperatura_f)
}
\arguments{
\item{temperatura_f}{{temperatura en grados fahrenheit}}
}
\value{
retorna el grado fahrenheit en centigrados
}
\description{
#Esta funcion te pide grados en fehrenheit para posteriormente devolvertelos en grados centigrados
}
\examples{
fahrenheit_a_centigrados(98.6)
#Esta funcion deberia devolver 37

}
