% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seguridad.R
\name{generar_grafo}
\alias{generar_grafo}
\title{Generar Grafo}
\usage{
generar_grafo(archivo_datos, separador = ",")
}
\arguments{
\item{archivo_datos}{ruta al archivo CSV que contiene la matriz de adyacencia.
La matriz tiene que tener cabecera tanto horizontal como vertical.}

\item{separador}{caracter que separa los campos en el CSV, por defecto ,}
}
\value{
el grafo correspondiente a la matriz de adyacencia.
}
\description{
Funci”n que genera un grafo a a partir de una matriz de adyacencia contenida
}
