% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mensaje_humedad.R
\name{mensaje_humedad}
\alias{mensaje_humedad}
\title{Mensaje humedad}
\usage{
mensaje_humedad(humedad_media)
}
\arguments{
\item{humedad_media}{{parametro que ingresa el usuario}}
}
\value{
Devuelve un mensaje en base a la humedad que hay
}
\description{
Mensaje humedad
}
\examples{
mensaje_humedad(30)

}
