% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obtener_temperatura_dia.R
\name{obtener_temperaturas_dia}
\alias{obtener_temperaturas_dia}
\title{Title}
\usage{
obtener_temperaturas_dia(data, fecha_input)
}
\arguments{
\item{data}{{dataset}}

\item{fecha_input}{{fecha de la que queremos saber la temp. max. y min.}}
}
\value{
Esta funcion nos retorna la temperatura maxima y minima de un dia en expecifico
}
\description{
Title
}
\examples{
obtener_temperaturas_dia(meteorologia_raw_data, "1973-01-01")

}
