% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pega_lista.R
\name{pega_lista}
\alias{pega_lista}
\title{Combina dataframes que están en listas (Combine dataframes that are in lists)}
\usage{
pega_lista(data, nc)
}
\arguments{
\item{data}{lista de dataframes}

\item{nc}{nombre de la columna que incluirá el nombre de la lista}
}
\value{
data frame
}
\description{
Primero agrega una columna al dataframe con el nombre de la lista, despues pega los dataframe
}
\examples{
lista1 <- list(a = data.frame(v = 1:3), b = data.frame(v = 4:6))
pega_lista(lista1, "nueva_columna")


}
