% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/punto_rocio.R
\name{punto_rocio}
\alias{punto_rocio}
\title{Punto rocio}
\usage{
punto_rocio(rocio_medio)
}
\arguments{
\item{rocio_medio}{{este es el argumento que el usuario ingresa}}
}
\value{
esta funcion te devuelve un mensaje sobre el punto de rocio
}
\description{
Punto rocio
}
\examples{
punto_rocio(3)

}
