% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_freq_columnas.R
\name{tabla_freq_columnas}
\alias{tabla_freq_columnas}
\title{Tabla de frecuencias con varias columnas (Frequency table with many columns)}
\usage{
tabla_freq_columnas(data, nomvar, peso = NULL, starts = NULL, na = NULL)
}
\arguments{
\item{data}{frame}

\item{nomvar}{vector con el nombre de la variable categórica (categorical variable)}

\item{peso}{en caso hayan pesos (if there are weights)}

\item{starts}{TRUE si se quiere aplicar sobre columnas que comparten los mismos carácteres al inicio}

\item{na}{en caso se consideren valores perdidos ()}
}
\value{
data frame con la cantidad de casos (n) y porcentaje (prop)
}
\description{
Aplica [tabla_freq()] a varias columnas
}
\examples{
tabla_freq_columnas(iris, "Sepal", starts = TRUE)

}
