% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_freq_estrato.R
\name{tabla_freq_estrato}
\alias{tabla_freq_estrato}
\title{Tabla de frecuencias segun estrato (Frequency table by group)}
\usage{
tabla_freq_estrato(data, var, grupo, peso = NULL, compara = NULL, na = NULL)
}
\arguments{
\item{data}{data frame}

\item{var}{variable categórica (categorical variable)}

\item{grupo}{grupo o estrato}

\item{peso}{en caso hayan pesos (if there are weights)}

\item{compara}{útil para comparar los porcentajes según el grupo}

\item{na}{en caso se consideren valores perdidos ()}
}
\value{
data frame con la cantidad de casos (n) y porcentaje (prop)
}
\description{
Aplica [tabla_freq()] según estrato
}
\examples{
tabla_freq_estrato(mtcars, var = "gear", grupo = "carb")


}
