% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor2.R
\name{cor2}
\alias{cor2}
\title{Correlaciones pairwise}
\usage{
cor2(data, round = 2)
}
\arguments{
\item{data}{Dataframe con los variables a correlacionar}

\item{round}{Redondear, por default = 2}
}
\value{
Matriz de correlaciones
}
\description{
Correlaciones pairwise
}
\examples{
cor2(mtcars)
}
