% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seguridad.R
\name{crear_grafo_desde_tabla_enlaces}
\alias{crear_grafo_desde_tabla_enlaces}
\title{Crear grafo a partir de lista de enlaces.}
\usage{
crear_grafo_desde_tabla_enlaces(fichero, identNumerico, sep = "")
}
\arguments{
\item{fichero}{ruta al archivo que contiene la lista de enlaces, puede ser .txt o .csv.
Cada fila del fichero contiene los dos v?rtices que forman cada enlace.}

\item{sep}{car?cter que separa en el fichero los dos v?rtices de cada enlace. Por defecto ser?
uno o más espacios en blanco}

\item{identificadores}{booleano que indica si el identificador de cada v?rtice es n?merico
o no. En caso de ser n?merico la funci?n lo convierte a car?cteres.}
}
\value{
el grafo correspondiente a la lista de enlaces.
}
\description{
Funci?n que genera un grafo a a partir de la lista de enlaces entre v?rtices.
}
