% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_freq.R
\name{tabla_freq}
\alias{tabla_freq}
\title{tabla freq}
\usage{
tabla_freq(data, x, peso = NULL)
}
\arguments{
\item{data}{Dataframe}

\item{x}{Variable categórica}

\item{peso}{En caso hayan pesos}
}
\value{
Dataframe
}
\description{
Lo mismo que table() pero devuelve un dataframe y acepta pesos
}
\examples{
tabla_freq(mtcars, "gear")


}
